SENTIMENTANALYSIS/requirements.txt
nltk==3.5
pymongo==3.10.1
python-dotenv==0.13.0
textblob==0.15.3
newspaper3k==0.2.8
dnspython==1.16.0